+++ src/commands/githubActions/bumpUpstream/github/index.ts
@@ -1,0 +1,6 @@
+export * from "./closeOldPrs.js";
+export * from "./getBumpPrBody.js";
+export * from "./getGithubSettings.js";
+export * from "./getUpstreamVersionTag.js";
+export * from "./isBranchNew.js";
+export * from "./isValidRelease.js";
