+++ src/commands/githubActions/bumpUpstream/github/isBranchNew.ts
@@ -1,0 +1,11 @@
+import { Github } from "../../../../providers/github/Github";
+import { getLocalBranchExists } from "../../../../utils/git";
+
+export async function isBranchNew({ branchName, repo }: { branchName: string, repo: Github }): Promise<boolean> {
+    const [remoteBranchExists, localBranchExists] = await Promise.all([
+        repo.branchExists(branchName),
+        getLocalBranchExists(branchName),
+    ]);
+
+    return !remoteBranchExists && !localBranchExists;
+}
